Split flux_perpare_map_data to align_map_data and combine_map_dataset.
