+++ fluxy/operators/flux_align_dataset.py
@@ -81,3 +81,43 @@ def align_lat_lon(
         )
 
     return aligned_ds_list
+
+
+def align_map_data(
+    ds_all: dict[xr.Dataset]
+) -> dict[xr.Dataset]:
+    """
+    Prepare flux datasets for flux maps by filtering variables, removing unused dimensions, and aligning time and spatial coordinates.
+
+    Args:
+        ds_all (dict[xr.Dataset]):
+            Dictionary of model names and corresponding xarray datasets.
+
+    Returns:
+        dict[xr.Dataset]:
+            Aligned datasets, after removing non-geographic variables.
+    """
+
+    for key, ds in ds_all.items():
+        # Step 1: Remove variables without 'time', 'latitude' and 'longitude'
+        ds = ds.drop_vars(
+            [
+                var
+                for var in ds.data_vars
+                if not {"time", "latitude", "longitude"}.issubset(ds[var].dims)
+            ]
+        )
+        # Step 2: Remove unused coordinates (dimensions that are no longer used)
+        unused_dims = set(ds.dims) - set(
+            dim for var_i in ds.data_vars for dim in ds[var_i].dims
+        )
+        ds_all[key] = ds.drop_dims(unused_dims)
+
+    # Align dataset coordinates
+    models = list(ds_all.keys())
+    ds_list = list(ds_all.values())
+    ds_list = align_time(ds_list)
+    ds_list = align_lat_lon(ds_list, coord="latitude")
+    ds_list = align_lat_lon(ds_list, coord="longitude")
+
+    return dict(zip(models, ds_list))
